requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
python-dotenv
google-generativeai #Gemini API
google-api-python-client #Google Search API